Solved BUG--Checkboxes have Strange Behavior
The bug was caused by the for loop in MainFrame.java that populated the
table with rows. The loop was giving the table pointers to the same two
rows, so all of the rows in the table were actually pointed to the same
objects.

hsTable and hsTableModel now only allow strings as column names.

Test code was added to solve the bug and needs to be removed for the
next commit.