Major Code Refactoring
The SkoonieFrame class has been created and implemented throughout most
of the program.

All of the material action frames are now subclasses of ActionFrame.
They also use the CancelConfirmPanel class for the cancel and confirm
buttons.

The SkoonieGuiUpdater (based on GuiUpdater by Mike S.) class has been
created in the same package as the SkoonieFrame class.

Function has been added to Tools class to generate an action command
from anywhere in the program. Eventually, this function should be used
for everything that needs to be caught in the Controller.