Program now Accesses Remote MySQL Database
The program now accesses a remote MySQL Database and reads data from it.

The MySQLDatabase class has been created to connect to and send and
retrieve data to and from the MySQL database containing all of the data
for the program. Currently, it has minimal functionality.

The Connector/J driver necessary for accessing a MySQL database has been
downloaded and copied into the project folder. The jar file was added to
the project as a library. Git has been told not to ignore the jar file.

Code for testing and debugging the MySQLDatabase class has been added to
the MainController class.

MainController class also now makes use of the Java Logger.