Customer Class Created to Store Info About a Customer
The Customer class was created to store information about a customer.

Testing code was added to the MainController class to make sure that we
are properly connecting to the database and retrieving the customer
information.

Some minor code refactoring in MySQLDatabase.

This commit is related to Asana tasks:
Create Customer Class to Hold Info About a Customer
Output All of the Customer(s) Information to the Console