Skoonie Key is Now Integrated into the Program
The Skoonie Key is now used as a "behind the scenes" way to reference
records in a table. Every table has a Skoonie Key column, and every
record has a Skoonie Key unique to that table.

The Skoonie Key should NEVER be changed.

Also, all classes that store information about a record in a table now
extend the Record class. This currently includes Batch and Customer.

The Rack class and everything to do with racks was removed from existence
because we don't need it yet; program still in early development stages.
removed from the table

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>