Command can now be initiated as now a wrapper class for messages
The command part of the CMVC now works more like an assembly line:
Once perform() is called on an instance of Command, the assembly line
begins in controller where information and data can be added to the
Command instance. The instance will then be passed all the way down the
assembly line, and each "station", or class, can choose to contribute his
part to a command and perform his own actions or do nothing and just send
it further down the line.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>