Combined inputs in ReceiveMaterialWindow
If any of the inputs for the Receivement and Batch descriptors have the
same name, then they are now represented by one input field. There are only
two exception: the Id for a batch and the Id for a receivement.

Also, the Date for the Receivement is manually added in as the third item.

This is just so that the input field appears where we want it to.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>